lambda/create_report/requirements-poetry.txt
boto3==1.34.62 ; python_version >= "3.10" and python_version < "4.0"
botocore==1.34.62 ; python_version >= "3.10" and python_version < "4.0"
calplot==0.1.7.5 ; python_version >= "3.10" and python_version < "4.0"
contourpy==1.2.0 ; python_version >= "3.10" and python_version < "4.0"
cycler==0.12.1 ; python_version >= "3.10" and python_version < "4.0"
defusedxml==0.7.1 ; python_version >= "3.10" and python_version < "4.0"
fonttools==4.49.0 ; python_version >= "3.10" and python_version < "4.0"
fpdf2==2.7.8 ; python_version >= "3.10" and python_version < "4.0"
jmespath==1.0.1 ; python_version >= "3.10" and python_version < "4.0"
kiwisolver==1.4.5 ; python_version >= "3.10" and python_version < "4.0"
matplotlib==3.8.3 ; python_version >= "3.10" and python_version < "4.0"
numpy==1.26.4 ; python_version >= "3.10" and python_version < "4.0"
packaging==24.0 ; python_version >= "3.10" and python_version < "4.0"
pandas==2.2.1 ; python_version >= "3.10" and python_version < "4.0"
pillow==10.2.0 ; python_version >= "3.10" and python_version < "4.0"
pyparsing==3.1.2 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.10" and python_version < "4.0"
pytz==2024.1 ; python_version >= "3.10" and python_version < "4.0"
s3transfer==0.10.0 ; python_version >= "3.10" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
tzdata==2024.1 ; python_version >= "3.10" and python_version < "4.0"
urllib3==2.0.7 ; python_version >= "3.10" and python_version < "4.0"
wordcloud==1.9.3 ; python_version >= "3.10" and python_version < "4.0"
